LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-rect-expected.txt


FAIL Test for checking position of the svg element when animateMotion's mpath is a rect assert_approx_equals: expected 100 +/- 1 but got 0
